+++ backend/src/main/kotlin/code/nebula/cipherquest/models/requests/CreateUserLevelRequest.kt
@@ -1,0 +1,6 @@
+package code.nebula.cipherquest.models.requests
+
+data class CreateUserLevelRequest(
+    var username: String,
+)
+
